feat(prayer-times,integration-demo): Implement prayer times calendar system with interactive demo
- Add prayer times calendar database migration (013_prayer_times_calendar_system.sql)
- Create comprehensive integration demo application with Cargo project structure
- Implement demo models for prayer times and calendar data structures
- Add demo services layer for prayer times calculations and calendar operations
- Create interactive web interface with static HTML for prayer times visualization
- Build complete demo application showcasing prayer times calendar integration
- Enable users to view, plan, and interact with Islamic prayer schedules
- Provide foundation for testing prayer times service across the platform

+++ src/ai_service/rag_system.rs
@@ -1331,7 +1331,7 @@ impl LLMInterface {
         }
         
         if weight_sum > 0.0 {
-            (total_confidence / weight_sum).min(1.0)
+            ((total_confidence / weight_sum) as f32).min(1.0)
         } else {
             0.5
         }
